change int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        int [] sales = {34, 54, 21, 76, 345, 643, 12, 456, 68};
+        long [] sales = {34, 54, 21, 76, 345, 643, 12, 456, 68};
         SalesManager salesManager = new SalesManager(sales);
 
         System.out.println(salesManager.max());
